requirements.txt
deepgram-sdk==5.3.0
Flask==3.1.0
Flask-Cors==5.0.0
python-dotenv==1.0.1
Werkzeug==3.1.3
